refactor: extract the domain layer
