+++ app.domain/Models/EventAggregate/Enums/EventHostingType.cs
