+++ app.domain/Models/EventAggregate/Enums/EventStatus.cs
