+++ app.domain/Models/EventAggregate/Enums/EventType.cs
