+++ app.domain/Models/EventAggregate/ValueObjects/ActivitySchedule.cs
