+++ app.domain/Models/EventAggregate/ValueObjects/EventSchedule.cs
