+++ app.domain/Models/EventAggregate/ValueObjects/Host.cs
