+++ app.domain/ViewModels/EventModel.cs
@@ -2,7 +2,7 @@ using app.Models.EventAggregate.Entities;
 using app.Models.EventAggregate.Enums;
 using app.Models.EventAggregate.ValueObjects;
 
-namespace app.ViewModels;
+namespace app.domain.ViewModels;
 public class EventModel
 {
     public string Title { get; set; } = string.Empty;
