+++ app/Components/Components/NewActivityPopup.razor
@@ -1,4 +1,5 @@
 @using app.Components.Pages.Admin.Events
+@using app.domain.ViewModels
 @inject IJSRuntime JsInterop
 
 <div id="activity-popup" class="activity-popup absolute top-0 left-0 w-full h-full z-10">
