+++ app/Components/_Imports.razor
@@ -21,4 +21,3 @@
 @using ErrorOr
 @using app.Models.EventAggregate.Enums
 @using app.Models.EventAggregate.ValueObjects
-@using app.ViewModels
