+++ app/Persistence/Repositories/Base/Repository.cs
@@ -1,5 +1,5 @@
 using System.Linq.Expressions;
-using app.Models.Common;
+using app.domain.Models.Common;
 using Microsoft.EntityFrameworkCore;
 
 namespace app.Persistence.Repositories.Base;
